requirements.txt
streamlit
PyPDF2
faiss-cpu
sentence-transformers
transformers
numpy
pytesseract
Pillow
requests
